fix(subtreevalidation): apply receive-side cap to CheckBlockSubtrees leaf-count gate
The peer-fetch fallback at check_block_subtrees.go bounds the HTTP body by
MaxIncomingSubtreeBytes but immediately re-validates the derived leaf count
against BlockAssembly.MaximumMerkleItemsPerSubtree. On the docker quickstart
profile (assembly cap 32k * 32 bytes = 1 MiB) a 2 MiB peer subtree passes the
new 128 MiB body cap and then fails the leaf-count gate with "exceeds policy
max" — the same stuck-at-height regression #905 set out to fix, one
statement later.

Derive the leaf-count bound from MaxIncomingSubtreeBytes / chainhash.HashSize.
The bounded HTTP read already enforces this implicitly; the explicit check
remains as a guard before subtreepkg.NewIncompleteTreeByLeafCount allocates
against the count.

Also:
- Change MaxIncomingSubtreeBytes from int to int64; it's a byte count and
  DoHTTPRequestBounded already takes int64. Drop the redundant int64(...)
  casts at the three call sites.
- Add TestCheckBlockSubtrees_LocalAssemblyPolicyIgnored mirroring the
  existing regression tests for fetchSubtreeFromPeer and getSubtreeTxHashes.

+++ services/blockvalidation/get_blocks.go
@@ -662,7 +662,7 @@ func (u *Server) fetchSubtreeFromPeer(ctx context.Context, subtreeHash *chainhas
 	// streams more than this is malicious — fail fast rather than ReadAll into memory.
 	// This must be independent of local BlockAssembly.MaximumMerkleItemsPerSubtree, which
 	// only controls what *this node* assembles; peers may legitimately produce larger subtrees.
-	maxSubtreeBytes := int64(u.settings.SubtreeValidation.MaxIncomingSubtreeBytes)
+	maxSubtreeBytes := u.settings.SubtreeValidation.MaxIncomingSubtreeBytes
 
 	// Use the existing HTTP utility to fetch subtree
 	subtreeBytes, err := util.DoHTTPRequestBounded(ctx, url, maxSubtreeBytes)
